Refactor hixl and hcomm channels
